[UPDATE] 更新token缓存redis 验证登陆 重新修改Member类

+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbCartMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbCart;
 
 public interface TbCartMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbCart record);
 
     int insertSelective(TbCart record);
 
-    TbCart selectByPrimaryKey(Integer id);
+    TbCart selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbCart record);
