+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbCategoryMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbCategory;
 
 public interface TbCategoryMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbCategory record);
 
     int insertSelective(TbCategory record);
 
-    TbCategory selectByPrimaryKey(Integer id);
+    TbCategory selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbCategory record);
