+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbExpressMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbExpress;
 
 public interface TbExpressMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbExpress record);
 
     int insertSelective(TbExpress record);
 
-    TbExpress selectByPrimaryKey(Integer id);
+    TbExpress selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbExpress record);
