+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbItemMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbItem;
 
 public interface TbItemMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbItem record);
 
     int insertSelective(TbItem record);
 
-    TbItem selectByPrimaryKey(Integer id);
+    TbItem selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbItem record);
