+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbMemberMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbMember;
 
 public interface TbMemberMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbMember record);
 
     int insertSelective(TbMember record);
 
-    TbMember selectByPrimaryKey(Integer id);
+    TbMember selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbMember record);
