+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbOrderItemMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbOrderItem;
 
 public interface TbOrderItemMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbOrderItem record);
 
     int insertSelective(TbOrderItem record);
 
-    TbOrderItem selectByPrimaryKey(Integer id);
+    TbOrderItem selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbOrderItem record);
