+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbPayinfoMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbPayinfo;
 
 public interface TbPayinfoMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbPayinfo record);
 
     int insertSelective(TbPayinfo record);
 
-    TbPayinfo selectByPrimaryKey(Integer id);
+    TbPayinfo selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbPayinfo record);
