+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbPermissionMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbPermission;
 
 public interface TbPermissionMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbPermission record);
 
     int insertSelective(TbPermission record);
 
-    TbPermission selectByPrimaryKey(Integer id);
+    TbPermission selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbPermission record);
