+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbRoleMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbRole;
 
 public interface TbRoleMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbRole record);
 
     int insertSelective(TbRole record);
 
-    TbRole selectByPrimaryKey(Integer id);
+    TbRole selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbRole record);
