+++ mkshop-manager/mkshop-manager-dao/src/main/java/com/nexus/manager/mapper/TbRolePermMapper.java
@@ -3,13 +3,13 @@ package com.nexus.manager.mapper;
 import com.nexus.manager.pojo.TbRolePerm;
 
 public interface TbRolePermMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(TbRolePerm record);
 
     int insertSelective(TbRolePerm record);
 
-    TbRolePerm selectByPrimaryKey(Integer id);
+    TbRolePerm selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(TbRolePerm record);
